Remove real aws key values
The aws keys that were being passed in are only required to exist
not actually be right. The aws sdk loads these when it loads,
but they're not actually used in the local environment.

Also removed the CONT_DIR env var,
as there is no need for it to be configurable.